Refactor the experimental torch environments into the experimental folder.
