+++ hydrogym/experimental/__init__.py
@@ -1,0 +1,6 @@
+from core_1DEnvs import OneDimEnv, PDESolverBase1D
+
+from .kuramoto_sivashinsky import Kuramoto_Sivashinsky
+from .burgers import Burgers
+
+__all__ = ["OneDimEnv", "PDESolverBase1D", "Kuramoto_Sivashinsky", "Burgers"]
