+++ hydrogym/experimental/burgers/__init__.py
