+++ hydrogym/experimental/burgers/flow.py
