+++ hydrogym/experimental/burgers/ks_init.tensor
