+++ hydrogym/experimental/kuramoto_sivashinsky/ks_init.tensor
